Added stubs to People in Dean

+++ app/models/people_of_delta.go
@@ -2,5 +2,5 @@ package models
 
 type PeopleOfDelta struct {
 	Name string
-	RollNo int32
+	RollNo int64
 }
